make invocation counter better accessible
