+++ test/KubeOps.Operator.Test/Controller/Operator/TestController.cs
@@ -5,22 +5,22 @@ namespace KubeOps.Operator.Test.Controller.Operator;
 
 public class TestController : IEntityController<V1IntegrationTestEntity>
 {
-    private readonly ControllerMockService _svc;
+    private readonly InvocationCounter<V1IntegrationTestEntity> _svc;
 
-    public TestController(ControllerMockService svc)
+    public TestController(InvocationCounter<V1IntegrationTestEntity> svc)
     {
         _svc = svc;
     }
 
     public Task ReconcileAsync(V1IntegrationTestEntity entity)
     {
-        _svc.Reconcile(entity);
+        _svc.Invocation(entity);
         return Task.CompletedTask;
     }
 
     public Task DeletedAsync(V1IntegrationTestEntity entity)
     {
-        _svc.Delete(entity);
+        _svc.Invocation(entity);
         return Task.CompletedTask;
     }
 }
